Changed to f16s for gpu

+++ src/gpu/pipeline/stages/shaders/copy.rs
@@ -3,13 +3,14 @@ vulkano_shaders::shader! {
     ty: "compute",
     src: r"
     #version 450
+    #extension  GL_EXT_shader_16bit_storage : require
 
     layout(set = 0, binding = 0) buffer Source {
-        float arr[];
+        float16_t arr[];
     } source;
 
     layout(set = 0, binding = 1) buffer Target {
-        float arr[];
+        float16_t arr[];
     } target;
 
     void main() {
